Fatal error displaying the site's error page
The default Joomla! error page tries to render the Search module.
It does so by calling ModuleHelper::renderModule which, in turn,
triggers the onRenderModule plugin event. We do trap that event
but we also try to check the Joomla! document type to avoid
interfering in non-HTML pages. However, Joomla! does NOT have a
document object when it's in an error page. This causes a fatal
PHP error which breaks the error page.

Solution: check if the document object is null and assume we are
inside an error page.